refactor(plan-cards): streamline plan card components and improve responsiveness
- Consolidated `PlanCards` and `PlanCardsMobile` into a more organized structure.
- Removed unused imports and components for cleaner code.
- Updated `SubscriptionPlans` and `Plans` components to utilize the new responsive layout.
- Enhanced mobile and larger screen handling with dedicated components for better user experience.
