+++ apps/web/app/(home)/(public)/containers/subscription-plans/subscription-plans.tsx
@@ -4,11 +4,7 @@ import { Title } from '@coobrastur/ui/components/data-display/title';
 import { Container } from '@coobrastur/ui/components/layouts/container';
 import { Section } from '@coobrastur/ui/components/layouts/section';
 
-import { PlanCards, PlanCardsMobile } from '@/shared/components/plan-cards';
-import {
-  ResponsiveLargerThan,
-  ResponsiveSmallerThan,
-} from '@/shared/components/responsive';
+import { PlanCards } from '@/shared/components/plan-cards/plan-cards';
 
 export const SubscriptionPlans = () => {
   return (
@@ -24,12 +20,7 @@ export const SubscriptionPlans = () => {
           </Text>
         </Heading>
       </Container>
-      <ResponsiveSmallerThan breakpoint="md">
-        <PlanCardsMobile />
-      </ResponsiveSmallerThan>
-      <ResponsiveLargerThan breakpoint="md">
-        <PlanCards />
-      </ResponsiveLargerThan>
+      <PlanCards />
     </Section>
   );
 };
