--- apps/web/shared/components/plan-cards/index.ts
@@ -1,2 +1,0 @@
-export * from './plan-cards';
-export * from './plan-cards-mobile';
